sprint 4: adds background for ingredient details; small imrovmns for token access
