+++ src/components/burger-ingredients/components/burger-ingredients-item/burger-ingredients-item.tsx
@@ -11,7 +11,7 @@ import {
 } from "../../../../redux/slices/ingredients";
 import { useDispatch, useSelector } from "react-redux";
 import { useDrag } from "react-dnd";
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
 import { ROUTES } from "../../../app/app";
 
 interface Props {
@@ -30,6 +30,7 @@ const BurgerIngredientsItem = (props: Props) => {
   const { selectedIngredients } = useSelector(selectIngredientsState);
   const dispatch = useDispatch();
   const [count, setCount] = useState<number | null>(null);
+  const location = useLocation();
 
   useEffect(() => {
     // we reset the counter every time selected ingredients are updated
@@ -54,7 +55,7 @@ const BurgerIngredientsItem = (props: Props) => {
       <Link
         to={{
           pathname: `${ROUTES.INGREDIENTS}/${props.ingredientDetails._id}`,
-          state: { prevPath: "/" },
+          state: { background: location },
         }}
       >
         <li
